refactor: Move all Exceptions defined in backup into a sealed BackupException class
relates #24465
