+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/es/backup/BackupManager.java
@@ -11,8 +11,8 @@ import io.camunda.tasklist.property.TasklistProperties;
 import io.camunda.tasklist.schema.backup.BackupPriority;
 import io.camunda.tasklist.schema.backup.Prio1Backup;
 import io.camunda.tasklist.schema.indices.IndexDescriptor;
+import io.camunda.webapps.backup.BackupException.GenericBackupException;
 import io.camunda.webapps.backup.BackupService;
-import io.camunda.webapps.backup.repository.BackupException;
 import java.util.Arrays;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,7 +28,7 @@ public abstract class BackupManager implements BackupService {
     if (index instanceof IndexDescriptor) {
       return ((IndexDescriptor) index).getFullQualifiedName();
     } else {
-      throw new cBackupException("Can't find out index name for backup.");
+      throw new GenericBackupException("Can't find out index name for backup.");
     }
   }
