--- webapps-backup/src/main/java/io/camunda/webapps/backup/exceptions/InvalidRequestException.java
@@ -1,14 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.webapps.backup.exceptions;
-
-public class InvalidRequestException extends RuntimeException {
-  public InvalidRequestException(final String message) {
-    super(message);
-  }
-}
