--- webapps-backup/src/main/java/io/camunda/webapps/backup/repository/GenericBackupException.java
@@ -1,18 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.webapps.backup.repository;
-
-public class GenericBackupException extends RuntimeException {
-  public GenericBackupException(final String message) {
-    super(message);
-  }
-
-  public GenericBackupException(final String message, final Exception cause) {
-    super(message, cause);
-  }
-}
